pyproject.toml
[tool.poetry]
name = "extend-euserv"
version = "0.1.0"
description = "Extend expiry date of euserv.de"
authors = ["freemt"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ffreemt/extend-euserv"

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.7.3"
logzero = "^1.6.3"
pyquery = "^1.4.3"
pyppeteer2 = "^0.2.2"
absl-py = "^0.11.0"
python-dotenv = "^0.15.0"
chardet = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.20.0"
flake8 = "^3.8.4"
black = "^20.8b1"
pep257 = "^0.7.0"

[tool.poetry.scripts]
extend-euserv = "extend_euserv.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
